Show prompt
A note on the use of `ExecuteProcess`: it is the simplest but also
most limited cross platform way of running an external command. It
is synchronous so it will stop Two Shell from taking new commands
before the current one has completed, which is what is required.
If a more control over the external process is required in the
future `TProcess` or `RunCommand` in the `process` unit are
suitable alternatives.