Accept multiple commands instead of exiting the program
To get user input I switched from using `Read()` to `Readln()`. When
using `Read()` inside the loop the program would not stop for user
input consistently on the second iteration of the loop. `input`
would be become an empty string which would be used to run the
external command, resulting in an error when no matching command is
found. Switching to use of `Readln()` solved this issue.

`Readln()` advances to the next line of the file after getting the
consuming line as input. In the context of this program the file is
standard input (stdin). Since `Read()` does not advance to the next
line of stdin when the loop returns to the second call of `Read()`
the program is still on the previous line of stdin, which has
already been consumed. Therefore the program does not stop for
input and an empty string is the result. `Readln()` advances to
the next line of stdin ready for the next loop iteration, fixing
this issue.

https://wiki.freepascal.org/Read