Handle errors when executing external commands
The compiler mode `OBJFPC` is required to use exceptions,
because of the use of classes.

On entering no command the shell will simply print the prompt
again and allow the user to enter another command. If there is an
error executing the command the shell will write to standard
output, and prompt for the next command.

There is some whitespace correction. The formatter isn't quite
getting it correct. I suspect the length of the main procedure
is confusing both visually and for the formatter. The problem may
clear up when the code is properly broken down into functions and
procedures later.