Pipes with TProcess signaling SIGPIPE broken pipe
First attemps at making pipes.

On line 123 the current process reads from stdout into the buffer.
This works, however when the next process writes to stdin from the
buffer a SIGPIPE signal is raised.

The current process is not running after it is read and the attempt
to write to the next process causes the broken pipe. See
`man 7 pipe` *I/O on pipes and FIFOs* for details.

This code follows the example from the Free Pascal wiki 'How to
redirect output with TProcess':
https://wiki.lazarus.freepascal.org/Executing_External_Programs#Using_input_and_output_of_a_TProcess

I cannot figure out why the current process ends or how to keep
its end of the pipe open for writing.

The next attempt will make use of the FCL Pipes unit, which has
`CreatePipeHandles` and `CreatePipeStreams`. This is possibly an
alternative to TProcess.