Swap TProcess Read with ReadBuffer
`ReadBuffer` does not cause the broken pipe issues that `Read`
did. Stepping through with a debugger shows that for the command
`ls -l | grep twosh` all the output from `ls` gets read into the
buffer and written to `grep`. The program finishes the pipe loop
and moves to the next command, which is `grep`.

Since `grep` is the only command left the program goes into the
single command (non-pipe) loop. When the debugger passes over the
`execute` procedure call for `currentCommand` the debugger stops
with a forked process.

The output from grep is never seen. The program is making a new
`grep` process that likely never gets anything from stdin and/or
receives to few arguments.

What needs to happen is at the end of the pipes loop after the
stdin for the next command is closed and the program has finished
waiting for it to execute, `nextCommand` should be assigned
to `currentCommand`. In the next iteration of the main loop the
program will have a `grep` process that already has input and needs
to emit its output.

Alternatively at the end of the pipes loop after `nextProcess` has
had its input closed, its output should be emitted to the shell.
Then the Program should increment its commands loop counter again
to skip over the already run `grep` process.