Start breaking code down into procedures and functions
I would have like to have had the code completed when compared
feature for feature against the original tutorial, but the current
code is too difficult to reason about.

The next step is to break the code down into procedures and
functions so it is easier to understand. After that the pipes
feature will then be implemented.