Prevent formatter inserting blank line after function declaration
The formatter was inserting a blank line before the `const` and
'var' keywords. My preference is for function bodies to begin
immediately after the declaration to visually reinforce that the
body belongs to the header.

The effect is that instead of this:
```
function MyFunction(): integer;

const
  default = 1;

begin
  exit(default)
end;
```

The formatter will permit this:
```
function MyFunction(): integer;
const
  default = 1;

begin
  exit(default)
end;
```