Finished creating functions and fixing compiler errors
The program is now factored out into enough functions and
procedures to be understandable, and compiler errors have been
fixed.

The program is in a broken state. It will run but no commands work.
The `exit` builtin still works however, so it is possible to start
and exit the shell.