Match the length of parsed commands to the length of the command line
Previously the length of the array holding the parsed commands was
set at 100, which was just an arbitrary upper limit chosen without
other reason. This meant that the parse commands array had a length
that would almost always be greater than its contents. The program
iterates over the entire length of 100 elements even if there are
only 0, 1, or 50 commands contained within.

Now the length of the parsed commands array should match the number
of commands split out from the standard input.

At this stage of debugging the builtin commands work. No other
commands work.