If no command is entered skip the current loop as soon as possible
The program splits the standard input on pipe `|` characters.
The way `SplitString()` works is that even if an empty string was
recieved from standard input the returned value will still have a
length of 1 with the only element being an empty string.

If the input is an empty string then `ReadLine()` will return an
empty `TStringDynArray`, which allows the program to easily check
if any commands were entered by checking for a 0 length.

Using `High()` to check that the string has a length of more than
0 has not been reliable, often returning 0. I might not be using it
correctly since it is intended to return the upper index of an
array. `Length()` has been more reliable for checking if a
collection is empty.