When using pipes skip the loop over both commands
Given the command `ls -l | grep twosh`, after running the piped
commands, on the next iteration of the loop the command `grep twosh`
would be run. This is because the loop moves through all the
commands one index at a time. This second command is a new
process independent of the initial piped command.

I've switched to a `while` loop to make this possible to skip over
the next iteration of the loop so a piped command will not run the
second command again. It is not possible to assign a new value to
a `for` loop counter variable.

This does limit the shell to piping only two commands in a
sequence. This is acceptable for now.