Execute piped commands
Two commands can now be piped together. The stdout of the first
command will be piped into the stdin of the second command. The
program is limited to piping only two commands.

At this stage the `twosh` program is feature complete when
compared to the original tutorial in the blog:

https://www.joshmcguigan.com/blog/build-your-own-shell-rust/

I referenced another example from the Free Pascal wiki: Executing
External Programs; Reading Large Output.

https://wiki.freepascal.org/Executing_External_Programs#Reading_large_output

To get the second commands output it is necessary to read from the
processes output into a buffer:

```
process.Output.Read(buffer, BUFFER_SIZE);
```

After that it is easy to create a subclass instance of TStream from
the buffer, which can be used to write to the `twosh` programs own
stdout.